add comment

+++ entrypoint.sh
@@ -1,12 +1,19 @@
 #!/bin/bash
 
-#git clone https://github.com/in-house-swagger/in-house-swagger.git
+# travis スクリプトの生成
 cd ~/target
 /home/travis/.travis/travis-build/bin/travis compile --no-interactive > ~/build.sh.tmp
-
-rm -R ~/build
-mkdir -p ~/build
-cp -r ~/target/* ~/build/
-cd ~/build
+# ビルド対象はgitからのチェックアウトではなく、ローカルリソースにするためコメントアウト
 cat ~/build.sh.tmp | sed -e "s/^travis_run_checkout/#travis_run_checkout/" > ~/build.sh
+
+# ビルド対象リソースの準備
+if [ -d ~/build ]; then
+    rm -R ~/build
+fi
+mkdir -p ~/build
+# リソースをコピー
+cp -r ~/target/* ~/build
+
+# ビルド実行
+cd ~/
 bash ~/build.sh
